+++ CHANGELOG.rst
@@ -1,6 +1,9 @@
 Unreleased
 ----------
 
+- Changes
+    - Ignore links and hashtags in fenced code blocks in Markdown.
+
 0.0.2 (2020-07-22)
 ------------------